requirements.txt
contourpy==1.0.7
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.3
mss==7.0.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
pyscreenshot==3.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
torch==1.13.1
typing_extensions==4.4.0
